separated old launch form classes (still used for gatelab) from new ones. Also renamed boutiquesTools bean classes
